refactor(dashboard): extract Mastio enrollments + badges sub-routers (F-B-201 PR-12/13)
Move the three /proxy/enrollments routes (approval queue: list +
approve + reject) plus the seven /proxy/badge/* HTMX status badges
(agents, enrollments, users, approvals, audit, updates, version)
out of the mcp_proxy/dashboard/router.py god-object into two
dedicated modules (enrollments_routes.py + badges_routes.py),
mounted via include_router.

Same pattern as PR-2..PR-11. The two clusters share zero helpers
(enrollments owns _resolve_agent_manager + _enroll_error_response +
ACTION_AGENT_ENROLL; badges owns the _license_mod feature gate),
so splitting matches the cohesion of every other sub-router.

F-B-201 progress: 12/13 PR. One left: core leftover (root + org-status
+ display-name + overview + settings + api/*).
